+++ components/commons/Button.tsx
@@ -1,0 +1,41 @@
+"use client";
+
+import React, { ButtonHTMLAttributes, ReactNode } from "react";
+import Image from "next/image";
+
+interface ButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
+  children: ReactNode;
+  isLoading?: boolean;
+}
+
+const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
+  ({ children, className = "", isLoading = false, ...props }, ref) => {
+    return (
+      <button
+        ref={ref}
+        className={`px-4 py-2 bg-blue-600 text-blue-500 rounded hover:bg-blue-900 transition ${className}`}
+        disabled={isLoading}
+        {...props}
+      >
+        {isLoading ? (
+          <div className="flex items-center gap-4">
+            <Image
+              src="/assets/icons/loader.svg"
+              alt="loader"
+              width={24}
+              height={24}
+              className="animate-spin"
+            />
+            Loading...
+          </div>
+        ) : (
+          children
+        )}
+      </button>
+    );
+  }
+);
+
+Button.displayName = "Button";
+
+export default Button;
