+++ components/commons/Footer.tsx
@@ -1,0 +1,9 @@
+const Footer: React.FC = () => {
+  return (
+    <footer className="p-4 bg-lime-500">
+      <p className="text-center font-bold text-light-200">&copy; 2024 Peer Voice. All rights reserved.</p>
+    </footer>
+  );
+};
+
+export default Footer;
