+++ components/forms/UserForm.tsx
@@ -5,8 +5,8 @@ import { z } from "zod";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useRouter } from "next/navigation";
 import { useState } from "react";
-import CustomFormField, { FormFieldType } from "../CustomFormField";
-import SubmitButton from "../SubmitButton";
+import CustomFormField, { FormFieldType } from "../commons/CustomFormField";
+import SubmitButton from "../commons/SubmitButton";
 import { Form } from "../ui/form";
 import { createUser } from "../../lib/actions/media_expert.actions";
 import { UserSchema } from "../../lib/validation";
