+++ lib/appwrite.config.ts
@@ -20,5 +20,4 @@ client.setEndpoint(ENDPOINT!).setProject(PROJECT_ID!).setKey(API_KEY!);
 
 export const databases = new sdk.Databases(client);
 export const users = new sdk.Users(client);
-export const messaging = new sdk.Messaging(client);
 export const storage = new sdk.Storage(client);
